refactor: reorganiza API-Ciudadano dejando solo el código del backend
