--- API-Ciudadano/.idea/encodings.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/SAFE-Rescue_Backend_API-Ciudadano/src/main/java" charset="UTF-8" />
-  </component>
-</project>
