--- API-Ciudadano/SAFE-Rescue_Backend_API-Ciudadano/src/main/java/com/SAFE_Rescue/API_Ciudadano/ApiAdministradorApplication.java
@@ -1,13 +1,0 @@
-package com.SAFE_Rescue.API_Ciudadano;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiAdministradorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApiAdministradorApplication.class, args);
-	}
-
-}
