--- API-Ciudadano/SAFE-Rescue_Backend_API-Ciudadano/src/main/java/com/SAFE_Rescue/API_Ciudadano/repository/CiudadanoRepository.java
@@ -1,14 +1,0 @@
-package com.SAFE_Rescue.API_Ciudadano.repository;
-
-import com.SAFE_Rescue.API_Ciudadano.modelo.Ciudadano;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CiudadanoRepository extends JpaRepository<Ciudadano, Long> {
-
-    public boolean existsByRun(Long run);
-
-    public boolean existsByTelefono(Long telefono);
-
-}
