--- API-Ciudadano/SAFE-Rescue_Backend_API-Ciudadano/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-spring.application.name=API_Ciudadano
-
-spring.datasource.url = jdbc:mysql://localhost:3306/db_SAFE_RESCUE
-spring.datasource.username=root
-spring.datasource.password=
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-
