+++ API-Ciudadano/src/main/java/com/SAFE_Rescue/API_Ciudadano/modelo/Login.java
@@ -1,0 +1,14 @@
+package com.SAFE_Rescue.API_Ciudadano.modelo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+public class Login {
+    private String correo;
+    private String contrasenia;
+
+}
