+++ API-Ciudadano/src/test/java/com/SAFE_Rescue/API_Ciudadano/ApiAdministradorApplicationTests.java
@@ -1,0 +1,13 @@
+package com.SAFE_Rescue.API_Ciudadano;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiAdministradorApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
